# api/openapi.yaml
openapi: 3.0.3
info:
  title: otakakotid
  description: otakakotid
  termsOfService: https://localhost:8787
  contact:
    email: kotaro.otaka@example.com
  license:
    name: MIT
  version: 0.0.1
externalDocs:
  description: OpenID Connect Core 1.0 incorporating errata set 1
  url: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html
servers:
  - url: http://localhost:7777
  - url: http://localhost:8888
tags:
  - name: Health
    description: Health check for the application
  - name: OIDC
    description: OpenID Connect
  - name: AuthN
    description: Authentication
  - name: WebAuthn
    description: Passkey authentication
paths:
  /:
    get:
      tags:
        - Health
      summary: Health Check
      description: Check the health status of the application
      operationId: healthCheck
      responses:
        "200":
          description: OK
        default:
          description: Internal Server Error
  /.well-known/openid-configuration:
    get:
      tags:
        - OIDC
      summary: OpenID Provider Configuration
      description: OpenID Provider Configuration
      operationId: openIDConfiguration
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenIDConfigurationResponseSchema"
        default:
          description: Internal Server Error
  /registration:
    description: send to email
    post:
      tags:
        - AuthN
      summary: Initialize Registration
      description: Initialize Registration
      operationId: initialiseRegistration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Registration successful"
        default:
          description: Internal Server Error
    get:
      tags:
        - AuthN
      summary: Finalize Registration
      description: Finalize Registration
      operationId: finalizeRegistration
      parameters:
        - name: code
          in: query
          description: code
          required: true
          schema:
            type: string
            example: 3RPHI3HA7DHB47LWWZESKFN67Q
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeAttestationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Registration successful"
        default:
          description: Internal Server Error
  /attestation:
    description: https://developer.mozilla.org/en-US/docs/Web/API/Web_Authentication_API/Attestation_and_Assertion#attestation
    get:
      tags:
        - WebAuthn
      summary: Initialize Attestation
      description: Initialize Attestation
      operationId: initializeAttestation
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitializeAttestationResponse'
          headers:
            Set-Cookie:
              description: Set-Cookie
              schema:
                type: string
        default:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - WebAuthn
      summary: Finalize Attestation
      description: Finalize Attestation
      operationId: finalizeAttestation
      parameters:
        - name: __attestation__
          in: cookie
          description: session
          required: true
          schema:
            type: string
            example: session
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeAttestationRequest'
      responses:
        '201':
          description: Created
        default:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /assertion:
    description: https://developer.mozilla.org/en-US/docs/Web/API/Web_Authentication_API/Attestation_and_Assertion#assertion
    get:
      tags:
        - WebAuthn
      summary: Initialize Assertion
      description: Initialize Assertion
      operationId: initializeAssertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitializeAssertionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitializeAssertionResponse'
          headers:
            Set-Cookie:
              description: Set-Cookie
              schema:
                type: string
        default:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - WebAuthn
      summary: Finalize Assertion
      description: Finalize Assertion
      operationId: finalizeAssertion
      parameters:
        - name: __assertion__
          in: cookie
          description: session
          required: true
          schema:
            type: string
            example: session
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeAssertionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeAssertionResponse'
        default:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OpenIDConfigurationResponseSchema:
      type: object
      description: https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderMetadata
      properties:
        issuer:
          type: string
          description: http://localhost:8787
          format: uri
          example: http://localhost:8787
        authorization_endpoint:
          type: string
          description: http://localhost:8787/authorize
          format: uri
          example: http://localhost:8787/authorize
        token_endpoint:
          type: string
          description: http://localhost:8787/token
          format: uri
          example: http://localhost:8787/token
        userinfo_endpoint:
          type: string
          description: http://localhost:8787/userinfo
          format: uri
          example: http://localhost:8787/userinfo
        jwks_uri:
          type: string
          description: http://localhost:8787/certs
          format: uri
          example: http://localhost:8787/certs
        revocation_endpoint:
          type: string
          description: http://localhost:8787/revoke
          format: uri
          example: http://localhost:8787/revoke
        subject_types_supported:
          type: array
          items:
            type: string
          example:
            - public
        id_token_signing_alg_values_supported:
          type: array
          items:
            type: string
          example:
            - RS256
      required:
        - issuer
        - authorization_endpoint
        - token_endpoint
        - userinfo_endpoint
        - jwks_uri
        - revocation_endpoint
        - subject_types_supported
        - id_token_signing_alg_values_supported
    RegistrationRequest:
      type: object
      description: Registration Request
      properties:
        email:
          type: string
          example: kotaro.otaka@example.com
      required:
        - email
    InitializeAttestationResponse:
      type: object
      description: Initialize Attestation Response
    InitializeAssertionResponse:
      type: object
      description: Initialize Assertion Response
    FinalizeAttestationRequest:
      type: object
      description: Finalize Attestation Request
    InitializeAssertionRequest:
      type: object
      description: Initialize Assertion Response
    FinalizeAssertionRequest:
      type: object
      description: Finalize Assertion Request
    FinalizeAssertionResponse:
      type: object
      description: Finalize Assertion Response
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message